requirements.txt
streamlit>=1.32
pyyaml>=6.0
